+++ CHANGELOG.md
@@ -1,3 +1,3 @@
 ## 0.0.1
 
-* TODO: Describe initial release.
+* 基于官方android-serialport-api项目，增加可配置奇偶校验、数据位、停止位；